Restore globals file [Read more]
Remove individual files
