--- js/state/schema.js
@@ -1,7 +1,0 @@
-export function setCurrentSchemaVersion(schemaVersion) {
-  window.sessionStorage.setItem("current-schema-version", schemaVersion);
-}
-
-export function getCurrentSchemaVersion() {
-  return window.sessionStorage.getItem("current-schema-version");
-}
